drevuz modified Yggdrasil/exemples/sortlist/builder.py

+++ exemples/sortlist/builder.py
@@ -38,6 +38,13 @@ if __name__ == "__main__":
             if key in dic and dic[key] == glob[key]:
                 del dic[key]
 
+    import os 
+    if not os.isfile("data.csv"):
+            msg = ("ce type d'exercice utilise un fichier de données csv : data.csv\n"
+               +"vous pouvez l'indiqer sous la forme suivante\n"
+               +"@ monfichier.csv [data.csv] ")
+        print(msg, file=sys.stderr)
+        sys.exit(1)
     dic['answer'],dic['sortlist']['items']= utils.builditemsandanswer()
 
     with open(output_json, "w+") as f:
